# api/openapi.yaml
openapi: 3.0.3
info:
  title: 直播借样 MVP API
  version: 1.0.0
  description: 钉钉借样助手一期 MVP 接口草案
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/returns/aggregations:
    get:
      tags: [Return]
      summary: 查询通用归还聚合清单
      operationId: queryReturnAggregations
      parameters:
        - in: query
          name: virtualStoreCode
          required: true
          schema:
            type: string
        - in: query
          name: sampleFilterType
          required: true
          schema:
            type: string
            enum: [ALL, PICKUP_ONLY]
      responses:
        '200':
          description: success
  /api/v1/returns/batches:
    post:
      tags: [Return]
      summary: 发起归还批次
      operationId: createReturnBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnBatchCreateRequest'
      responses:
        '200':
          description: success
  /api/v1/returns/batches/{returnBatchNo}:
    get:
      tags: [Return]
      summary: 查询归还批次详情
      operationId: getReturnBatchDetail
      parameters:
        - in: path
          name: returnBatchNo
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
  /api/v1/returns/batches/{returnBatchNo}/logistics:
    post:
      tags: [Return]
      summary: 填写归还物流
      operationId: fillReturnLogistics
      parameters:
        - in: path
          name: returnBatchNo
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnLogisticsRequest'
      responses:
        '200':
          description: success
...
components:
  schemas:
...
    ReturnBatchCreateRequest:
      type: object
      required:
        - virtualStoreCode
        - sourceType
        - sampleFilterType
        - returnMethod
        - items
      properties:
        virtualStoreCode:
          type: string
        sourceType:
          type: string
          enum: [TASK_DETAIL, GENERAL_RETURN]
        sampleFilterType:
          type: string
          enum: [ALL, PICKUP_ONLY]
        returnMethod:
          type: string
          enum: [EXPRESS, IN_PERSON]
        remark:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReturnBatchItem'
    ReturnBatchItem:
      type: object
      required:
        - skuCode
        - sizeCode
        - applyReturnQty
      properties:
        skuCode:
          type: string
        sizeCode:
          type: string
        applyReturnQty:
          type: integer
    ReturnLogisticsRequest:
      type: object
      required:
        - companyCode
        - companyName
        - logisticsNo
      properties:
        companyCode:
          type: string
        companyName:
          type: string
        logisticsNo:
          type: string